refactor: Tag naming change ActionTag(#26)
