+++ src/components/Tag/ActionTag.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryFn } from '@storybook/react'
 
-import Tag from './Tag'
+import Tag from './ActionTag'
 
 export default {
   title: 'Components/Tag',
